Automate node addition/terminatoin in the test run

+++ src/main/java/ConsensusTest/Tester.java
@@ -15,15 +15,18 @@ public class Tester {
 
     private final String kafkaServerAddress;
     private final Context jsContext;
+    private final int maxProcessCount;
     private String immortalProcess;
     private KafkaConsumer<String, String> kafkaConsumer;
     private String jarLocation;
     private final String kafkaTopic;
     private boolean terminate;
     private HashMap<String, Process> activeProcesses;
+    private boolean maxProcessCountReached;
 
 
-    public Tester(String jarLocation, String kafkaServerAddress, String kafkaTopic){
+
+    public Tester(String jarLocation, String kafkaServerAddress, String kafkaTopic, int maxProcessCount){
         this.jarLocation = jarLocation;
         this.kafkaTopic = kafkaTopic;
         this.kafkaServerAddress = kafkaServerAddress;
@@ -32,6 +35,8 @@ public class Tester {
         this.immortalProcess = null;
         this.activeProcesses = new HashMap<String, Process>();
         this.terminate = false;
+        this.maxProcessCountReached = false;
+        this.maxProcessCount = maxProcessCount;
     }
 
     public void read(){
@@ -71,7 +76,7 @@ public class Tester {
     }
 
     public void startNewProcess(String jarLocation, String kafkaServerAddress, String kafkaTopic){
-        if (this.activeProcesses.size() < 26){
+        if (this.activeProcesses.size() < maxProcessCount){
             String nodeId = UUID.randomUUID().toString();
             System.setProperty("id", nodeId);
             System.out.println("Id of the new process : " + nodeId);
@@ -104,21 +109,44 @@ public class Tester {
             processToBeKilled.destroy();
             activeProcesses.remove(nodeId);
             System.out.println("Processes after killing one : " + activeProcesses);
+            if (this.activeProcesses.size() == 0){
+                //FINISH TESTING
+                this.terminate = true;
+            }
         }
     }
 
     public static void main(String args[]){
-        Tester tester = new Tester(args[0], args[1], args[2]);
+        Tester tester = new Tester(args[0], args[1], args[2], Integer.parseInt(args[3]));
         tester.read();
 
-        tester.startNewProcess(tester.jarLocation, tester.kafkaServerAddress, tester.kafkaTopic);
-        try {
-            Thread.sleep(8000);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
+        while(!tester.terminate){
+
+            int randInt = (int)(1 + Math.random()*tester.maxProcessCount);
+
+            if(randInt> tester.activeProcesses.size()){
+                if(!tester.maxProcessCountReached){
+                    tester.startNewProcess(tester.jarLocation, tester.kafkaServerAddress, tester.kafkaTopic);
+                    if( tester.activeProcesses.size() == tester.maxProcessCount){
+                        tester.maxProcessCountReached = true;
+                    }
+                }
+            }
+            else{
+                tester.killProcess();
+            }
+
+            int randWait = (int)(1 + Math.random()*8)*1000;
+
+            try {
+                Thread.sleep(randWait);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
         }
-        tester.killProcess();
-        tester.terminate = true;
+
+
+
 
     }
 }
